requirements.txt
streamlit==1.28.0
pandas==2.1.1
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
openpyxl==3.1.2  
plotly==5.17.0   